+++ lib/provider/nextcloud_auth_provider.dart
@@ -57,9 +57,9 @@ class NextcloudAuthProvider with ChangeNotifier {
   }
 
   Future<void> setCredentials(String urlNcResponse) async {
-    urlNcResponse= Uri.decodeFull(urlNcResponse);
+    urlNcResponse = Uri.decodeFull(urlNcResponse);
     final match = _expCred.firstMatch(urlNcResponse);
-    final user = match.group(2);
+    final user = match.group(2).replaceAll('+', ' ');
     final password = match.group(3);
     if (user != null && password != null) {
       await _setCredentials(