+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - First major release
 - Bugfixes
 - New webview plugin for NC login
+- Add french translation
 
 ## 0.9.12
 - Dark Theme & Amoled Mode